multiverso/requirements.txt
streamlit
matplotlib
networkx
pillow